Allow users to set the location of the datastore.
Allow uses to set the location of the datastore, this involves a
major API change.

Instead of returning a DatastoreManager object, the exports.datastoremanager
is now a function, which will create the datastore manager in the native layer.
This function returns a Q style promise with will provide the datastore or
error on completion.

fixes #6