Replaced underscore with lodash
Replaced the underscore dependency with a few smaller
lodash modules for the functions being used
isempty
isfunction
isobject
isstring

lodash.isarray is deprecated in favour of Array.isArray so
that is used as well.

Defined the lodash dependencies in package.json and plugin.xml.
Created lodash_funcs.js to export the necessary methods.
Updated plugin.xml and the require statements to use new
cloudant-sync.lodash_funcs.